network/websockets: extract authQueue
